separated tkf into multiple files/modules
